--- a/base57.go
+++ b/base57.go
@@ -1,11 +1,7 @@
 package shortuuid
 
 import (
-	"fmt"
-	"math"
-	"math/big"
-	"strings"
-
+	"encoding/binary"
 	"github.com/google/uuid"
 )
 
@@ -14,46 +10,43 @@ type base57 struct {
 	alphabet alphabet
 }
 
+const strLen = 22
+
 // Encode encodes uuid.UUID into a string using the most significant bits (MSB)
 // first according to the alphabet.
 func (b base57) Encode(u uuid.UUID) string {
-	var num big.Int
-	num.SetString(strings.Replace(u.String(), "-", "", 4), 16)
+	num := uint128{
+		binary.BigEndian.Uint64(u[8:]),
+		binary.BigEndian.Uint64(u[:8]),
+	}
 
-	// Calculate encoded length.
-	length := math.Ceil(math.Log(math.Pow(2, 128)) / math.Log(float64(b.alphabet.Length())))
-
-	return b.numToString(&num, int(length))
+	return b.numToString(num)
 }
 
 // Decode decodes a string according to the alphabet into a uuid.UUID. If s is
 // too short, its most significant bits (MSB) will be padded with 0 (zero).
 func (b base57) Decode(u string) (uuid.UUID, error) {
-	str, err := b.stringToNum(u)
+	buf, err := b.stringToNumBytes(u)
 	if err != nil {
 		return uuid.Nil, err
 	}
-	return uuid.Parse(str)
+	return uuid.FromBytes(buf)
 }
 
 // numToString converts a number a string using the given alphabet.
-func (b *base57) numToString(number *big.Int, padToLen int) string {
-	var (
-		out   []rune
-		digit *big.Int
-	)
+func (b *base57) numToString(number uint128) string {
+	var digit uint64
+	out := make([]rune, strLen)
 
-	alphaLen := big.NewInt(b.alphabet.Length())
-
-	zero := new(big.Int)
-	for number.Cmp(zero) > 0 {
-		number, digit = new(big.Int).DivMod(number, alphaLen, new(big.Int))
-		out = append(out, b.alphabet.chars[digit.Int64()])
+	i := 0
+	for number.Hi > 0 || number.Lo > 0 {
+		number, digit = number.quoRem64(uint64(b.alphabet.Length()))
+		out[i] = b.alphabet.chars[digit]
+		i++
 	}
-
-	if padToLen > 0 {
-		remainder := math.Max(float64(padToLen-len(out)), 0)
-		out = append(out, []rune(strings.Repeat(string(b.alphabet.chars[0]), int(remainder)))...)
+	for i < strLen {
+		out[i] = b.alphabet.chars[0]
+		i++
 	}
 
 	reverse(out)
@@ -61,31 +54,39 @@ func (b *base57) numToString(number *big.Int, padToLen int) string {
 	return string(out)
 }
 
-// stringToNum converts a string a number using the given alphabet.
-func (b *base57) stringToNum(s string) (string, error) {
-	n := big.NewInt(0)
+// stringToNumBytes converts a string a number using the given alphabet.
+func (b *base57) stringToNumBytes(s string) ([]byte, error) {
+	var (
+		n     uint128
+		err   error
+		index int64
+	)
 
 	for _, char := range s {
-		n.Mul(n, big.NewInt(b.alphabet.Length()))
-
-		index, err := b.alphabet.Index(char)
+		n, err = n.mul64(uint64(b.alphabet.Length()))
 		if err != nil {
-			return "", err
+			return nil, err
 		}
 
-		n.Add(n, big.NewInt(index))
-	}
+		index, err = b.alphabet.Index(char)
+		if err != nil {
+			return nil, err
+		}
 
-	if n.BitLen() > 128 {
-		return "", fmt.Errorf("number is out of range (need a 128-bit value)")
+		n, err = n.add64(uint64(index))
+		if err != nil {
+			return nil, err
+		}
 	}
-
-	return fmt.Sprintf("%032x", n), nil
+	buf := make([]byte, 16)
+	n.putBytes(buf)
+	return buf, nil
 }
 
 // reverse reverses a inline.
 func reverse(a []rune) {
-	for i, j := 0, len(a)-1; i < j; i, j = i+1, j-1 {
-		a[i], a[j] = a[j], a[i]
+	n := len(a)
+	for i := 0; i < n/2; i++ {
+		a[i], a[n-1-i] = a[n-1-i], a[i]
 	}
 }