v4.0.0:
```
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
BenchmarkUUID-12                  355501              3152 ns/op            2650 B/op         98 allocs/op
BenchmarkEncoding-12              414798              2828 ns/op            2640 B/op         98 allocs/op
BenchmarkDecoding-12             2422132               470.2 ns/op           152 B/op          5 allocs/op
```
with these optimizations:
```cpu: 12th Gen Intel(R) Core(TM) i5-12400F
BenchmarkUUID-12                 3777259               314.0 ns/op            48 B/op          2 allocs/op
BenchmarkEncoding-12             6792055               184.5 ns/op            32 B/op          1 allocs/op
BenchmarkDecoding-12             5202288               230.9 ns/op            64 B/op          2 allocs/op
```