@anatoly-kussul 2024-11-29
fixed decoding benchmark, as it was returning error, and not fully decoding.

last commit benchmarks:
```
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
BenchmarkUUID-12                 4058689               283.6 ns/op            40 B/op          2 allocs/op
BenchmarkEncoding-12             8116075               149.6 ns/op            24 B/op          1 allocs/op
BenchmarkDecoding-12             4644693               258.5 ns/op             0 B/op          0 allocs/op
```
v4.0.0 benchmarks with fixed decoding test:
```
cpu: 12th Gen Intel(R) Core(TM) i5-12400F
BenchmarkUUID-12                  363300              3179 ns/op            2650 B/op         98 allocs/op
BenchmarkEncoding-12              431427              2851 ns/op            2640 B/op         98 allocs/op
BenchmarkDecoding-12             1000000              1005 ns/op             192 B/op          8 allocs/op
 ```