@lithammer 2024-11-29
Very impressing work! 👏 You've managed to _drastically_ reduce the number of allocations. Here's the output [benchstat](https://pkg.go.dev/golang.org/x/perf/cmd/benchstat) which shows the difference in percentage.

Nice find out error in the benchmark test case as well.

```
goos: darwin
goarch: arm64
pkg: github.com/lithammer/shortuuid/v4
cpu: Apple M2
           │   old.txt    │               new.txt               │
           │    sec/op    │   sec/op     vs base                │
UUID-8       2782.5n ± 1%   479.6n ± 0%  -82.77% (p=0.000 n=10)
Encoding-8   2438.5n ± 2%   216.5n ± 2%  -91.12% (p=0.000 n=10)
Decoding-8    368.6n ± 0%   266.9n ± 0%  -27.58% (p=0.000 n=10)
geomean       1.357µ        302.6n       -77.71%

           │   old.txt    │                 new.txt                 │
           │     B/op     │    B/op     vs base                     │
UUID-8       2650.00 ± 0%   40.00 ± 0%   -98.49% (p=0.000 n=10)
Encoding-8   2640.00 ± 5%   24.00 ± 0%   -99.09% (p=0.000 n=10)
Decoding-8     152.0 ± 0%     0.0 ± 0%  -100.00% (p=0.000 n=10)
geomean       1020.7                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean

           │   old.txt   │                 new.txt                 │
           │  allocs/op  │ allocs/op   vs base                     │
UUID-8       98.000 ± 0%   2.000 ± 0%   -97.96% (p=0.000 n=10)
Encoding-8   98.000 ± 4%   1.000 ± 0%   -98.98% (p=0.000 n=10)
Decoding-8    5.000 ± 0%   0.000 ± 0%  -100.00% (p=0.000 n=10)
geomean       36.35                    ?                       ¹ ²
¹ summaries must be >0 to compute geomean
² ratios must be >0 to compute geomean
```