@lithammer 2024-11-29
Included in [v4.1.0](https://github.com/lithammer/shortuuid/releases/tag/v4.1.0). Thanks for your contribution!